feat(aws/local): transformReconcileNews hook + shared image-source trigger
- DevWatchProvider gains `transformReconcileNews`: transform the resolved
  props before EVERY delegated reconcile (engine- and watcher-driven),
  inside the per-id lock with the override services provided — the seam a
  dev provider uses to swap a build-here artifact for a build-locally one.
  The untransformed props stay in the watch registry.
- `imageSourceTrigger` moves from EcsDevWatch to the shared
  AWS/Local/ImageSourceTrigger module (re-exported for its ECS consumers)
  so the MicroVM dev provider can reuse it.
